Docs and move of materials
